multi/refactor: move forward manager into its own module
To introduce bidirectional reputation, we want to more cleanly separate
outgoing reputation from the top level forward manager.
